refactor(routes-cleanup): refactor routes into separate files for better readability
